Extracting logic to a Hook
