+++ tracks/src/hooks/useLocation.js
@@ -1,0 +1,32 @@
+import { useState, useEffect } from 'react';
+import {
+  Accuracy,
+  requestPermissionsAsync,
+  watchPositionAsync
+} from 'expo-location';
+
+export default callback => {
+  const [err, setErr] = useState(null);
+
+  const startWatching = async () => {
+    try {
+      await requestPermissionsAsync();
+      await watchPositionAsync(
+        {
+          accuracy: Accuracy.BestForNavigation,
+          timeInterval: 1000,
+          distanceInterval: 10
+        },
+        callback
+      );
+    } catch (e) {
+      setErr(e);
+    }
+  };
+
+  useEffect(() => {
+    startWatching();
+  }, []);
+
+  return [err];
+};
