+++ tracks/src/screens/TrackCreateScreen.js
@@ -1,41 +1,15 @@
 // import '../_mockLocation';
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useContext } from 'react';
 import { StyleSheet } from 'react-native';
 import { Text } from 'react-native-elements';
 import { SafeAreaView } from 'react-navigation';
-import {
-  requestPermissionsAsync,
-  watchPositionAsync,
-  Accuracy
-} from 'expo-location';
 import Map from '../components/Map';
 import { Context as LocationContext } from '../context/LocationContext';
+import useLocation from '../hooks/useLocation';
 
 const TrackCreateScreen = () => {
   const { addLocation } = useContext(LocationContext);
-  const [err, setErr] = useState(null);
-
-  const startWatching = async () => {
-    try {
-      await requestPermissionsAsync();
-      await watchPositionAsync(
-        {
-          accuracy: Accuracy.BestForNavigation,
-          timeInterval: 1000,
-          distanceInterval: 10
-        },
-        location => {
-          addLocation(location);
-        }
-      );
-    } catch (e) {
-      setErr(e);
-    }
-  };
-
-  useEffect(() => {
-    startWatching();
-  }, []);
+  const [err] = useLocation(addLocation);
 
   return (
     <SafeAreaView forceInset={{ top: 'always' }}>
